Design Refinements and Initial Signed Division
Refactored part of the code. Moved around the code for the simd operations
to be in a seperate file. The constant divider class now inherits from an
implementatiton class that determines whether the integer provided is signed
or not. The interface facing class forwards to the bases and contains some
common information.

Additionally, began implementatiton for signed division. The code has not been
tested though so teh signed part is not bug free.